+++ CloudShell.ResourceDefinitions.ResourceManager/ResourceModelResourceManagerStateProvider.cs
@@ -1,0 +1,9 @@
+using ResourceManagerState = CloudShell.Abstractions.ResourceManager.ResourceState;
+using ResourceModelResource = CloudShell.ResourceDefinitions.Resource;
+
+namespace CloudShell.ResourceDefinitions.ResourceManager;
+
+public interface IResourceModelResourceManagerStateProvider
+{
+    ResourceManagerState? GetState(ResourceModelResource resource);
+}